backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
httpx==0.28.1
openai==1.57.4
pillow==11.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
redis==5.2.1
numpy==2.2.1
torch==2.1.0
python-dotenv
anthropic
